% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFAII_Rook-data.R
\docType{data}
\name{HFAII_Rook}
\alias{HFAII_Rook}
\title{HFAII Rook Adjacency Matrix}
\format{This adjacency matrix is formated using rook neighbor criteria, meaning two locations on
 the visual field are only considered neighbors if they share an edge. The adjacency matrix is a
 54 x 54 dimensional binary object with zeros on the diagonal and the column and row sums are equal
 to the number of neighbors.}
\usage{
data(HFAII_Rook)
}
\description{
Binary adjacency matrix for the Humphrey Field Analyzer-II (Carl Zeiss Meditec Inc., Dublin, CA)
}
\keyword{datasets}
