% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_PosteriorWij.R
\name{is.PosteriorAdj}
\alias{is.PosteriorAdj}
\title{is.PosteriorAdj}
\usage{
is.PosteriorAdj(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.PosteriorAdj} is a general test of an object being interpretable as a
\code{\link{PosteriorAdj}} object.
}
\details{
The \code{\link{PosteriorAdj}} class is defined as the posterior adjacency
 object that results from the \code{\link{PosteriorAdj}} function.
}
