% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.STBDwDM}
\alias{is.STBDwDM}
\title{is.STBDwDM}
\usage{
is.STBDwDM(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.STBDwDM} is a general test of an object being interpretable as a
\code{\link{STBDwDM}} object.
}
\details{
The \code{\link{STBDwDM}} class is defined as the regression object that
 results from the \code{\link{STBDwDM}} regression function.
}
