% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_Adjacency.R
\name{PlotAdjacency}
\alias{PlotAdjacency}
\title{PlotAdjacency}
\usage{
PlotAdjacency(WijAlpha, Visit = 1, stat = "mean",
  main = "Estimated Adjacencies", color.scheme = c("Black", "White"),
  edgewidth = 2, cornerwidth = 1/4, lwd.border = 3, color.bs = "gray",
  zlim = c(0, 1), legend = TRUE, Degree = NULL)
}
\arguments{
\item{WijAlpha}{variable to be plotted on the visual field (e.g. differential light sensitivity).}

\item{Visit}{an overall title for the plot.}

\item{stat}{a label for the legend (default = "DLS (dB)").}

\item{main}{the limits used for the legend (default are the minimum and maximum of Y).}

\item{color.scheme}{the number of bins used to refine the color palette for the figure and legend.}

\item{edgewidth}{logical, indicating whether there should be a border around the visual field (default = TRUE).}

\item{cornerwidth}{logical, indicating whether the legend should be present (default = TRUE).}

\item{lwd.border}{a vector of character strings representing the color palette.}

\item{color.bs}{a vector of character strings representing the color palette.}

\item{zlim}{a vector of character strings representing the color palette.}

\item{legend}{a vector of character strings representing the color palette.}

\item{Degree}{a vector of character strings representing the color palette.}
}
\description{
Plots a heat map of the differential light sensitivity on the Humphrey Field
Analyzer-II visual field.
}
\details{
\code{PlotAdjacency} is used in the application of glaucaom progression to
 plot a variable across the visual field in the form of a heat map.
}
\examples{
\dontrun{
data(VFSeries)
PlotAdjacency(Y = VFSeries$DLS[VFSeries$Visit == 1],
                  main = "Sensitivity estimate (dB) at each \\n location on visual field",
                  legend.lab = "DLS (dB)",
                  zlim = c(10, 35),
                  bins = 250)
}

}
\author{
Samuel I. Berchuck
}

