% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womblR-package.R
\docType{package}
\name{womblR}
\alias{womblR}
\alias{womblR-package}
\title{womblR}
\description{
Currently this package implements a spatiotemporal boundary detection
with a dissimilarity metric for areal data with inference in a Bayesian setting
using Markov chain Monte Carlo (MCMC). The response variable can be modeled as
Gaussian (no nugget), probit or Tobit link and spatial correlation is introduced
at each time point through a conditional autoregressive (CAR) prior. Temporal
correlation is introduced through a hierarchical structure and can be specified as
exponential or first-order autoregressive. Full details of the the package can be found
in the accompanying vignette.
}
\author{
Samuel I. Berchuck \email{berchuck@ad.unc.edu}
}

